+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - [#1340](https://github.com/influxdata/telegraf/issues/1340): statsd: do not log every dropped metric.
+- [#1368](https://github.com/influxdata/telegraf/pull/1368): Add precision rounding to all metrics on collection.
 
 ### Bugfixes